create popup from posts
Each post will make a popup appear that will cover the whole window and display details about the post. This popup appears on tap of the expand or photo, and not the entire wrapper as if it was only for wrapper and you were on mobile, you couldn't read more without opening the entire photo.

In mobile the popup only shows the photo.